+++ HAPI_Start/src/Level.h
@@ -21,7 +21,7 @@ class Level
 public:
 	static std::unique_ptr<Level> loadLevel(const std::string& levelName, eGameDifficulty gameDifficulty);
 	
-	void addTurretAtPosition(Vector2i position, eTurretType turretType, int& playerScore);
+	void addTurretAtPosition(Vector2f position, eTurretType turretType, int& playerScore);
 	
 	bool isEnded() const;
