+++ HAPI_Start/src/Sprite.h
@@ -9,22 +9,22 @@ class Sprite
 {
 public:
 	Sprite();
-	Sprite(Vector2i startingPosition, int tileID);
+	Sprite(Vector2f startingPosition, int tileID);
 
 	int getID() const;
 	Vector2i getSize() const;
-	Vector2i getPosition() const;
+	Vector2f getPosition() const;
 	bool isAlpha() const;
 	Frame getFrame() const;
 	Rectangle getFrameRect() const;
 	Texture& getTexture() const;
 
 	void setID(int tileID);
-	void setPosition(Vector2i position);
+	void setPosition(Vector2f position);
 
 private:
 	std::reference_wrapper<Texture> m_texture;
-	Vector2i m_position;
+	Vector2f m_position;
 	Vector2i m_size;
 	int m_tileID;
 };
