+++ HAPI_Start/src/Utilities/Math.h
@@ -5,6 +5,6 @@
 
 namespace Math
 {
-	Vector2f getDirection(Vector2i currentPosition, Vector2i destination);
-	bool isWithinRange(Vector2i currentPosition, Vector2i targetPosition, float attackRange);
+	Vector2f getDirection(Vector2f currentPosition, Vector2f destination);
+	bool isWithinRange(Vector2f currentPosition, Vector2f targetPosition, float attackRange);
 }
